Generalized piecharts + average
Both pie charts and averages are generalized. Login window is much small as well, and the password shows as asterisks now. Changed the names of the stats file to the name of saved diagram files, since it's inside the stats folder anyways. For the program to run, you need to save a sequence diagram as "seq1", and a class diagram as "class1".